Migrated layout of PackedBubbleSeries to class pattern.
